warnings fix
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@29161 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tad/snmp/tad_snmp_layer.c
@@ -558,9 +558,10 @@ tad_snmp_match_bin_cb(csap_p           csap,
         if (0)
         {
             char buf[100], *p = buf;
-            int i;
+            unsigned int i;
+
             for (i = 0; i < vars->name_length; i++)
-                p += sprintf(p, ".%d", vars->name[i]);
+                p += sprintf(p, ".%d", (int)vars->name[i]);
             VERB("%s(): varbind oid: %s", __FUNCTION__, buf);
         }
         VERB("%s(): varbind type: %d, len %d", __FUNCTION__, vars->type, vars->val_len);
